requirements.txt
fastapi==0.115.12
pydantic==2.11.3
pymongo==4.12.0
python-dotenv==1.1.0
Requests==2.32.3
uvicorn==0.34.0